refactor GetPageDto and FindOneDto to common dto
