+++ src/abstract-media/abstract-media.dto.ts
@@ -1,22 +1,5 @@
-import { IntersectionType } from '@nestjs/swagger';
-import { Lang, LangQueryDto, MultiLang, PagedDto } from '../common.dto';
+import { MultiLang } from '../common.dto';
 import { Image, Audio } from '@luomus/laji-schema';
-import { ParseOptionalBoolean } from '../serializing/serializing';
-import { IsBoolean } from 'class-validator';
-
-export class GetPageDto extends IntersectionType(PagedDto, LangQueryDto) {
-    /**
-     * Comma separated ids
-     */
-    idIn?: string;
-}
-
-export class FindOneDto  {
-    lang?: Lang = Lang.en;
-    @ParseOptionalBoolean()
-    @IsBoolean()
-    langFallback?: boolean = true;
-}
 
 export enum MediaType {
     image = 'MM.image',
